Typecast dynamic promises to int (resolves #75)
resolves #75

Toil cannot handle floats in scientific notation for memory and disk, and really, fractions of ram
or disk are not logical so we shouldn't be using them anyway.

+++ src/protect/alignment/common.py
@@ -21,7 +21,7 @@ import os
 
 # disk for indexing
 def index_disk(bamfile):
-    return ceil(bamfile.size + 524288)
+    return int(ceil(bamfile.size + 524288))
 
 
 def index_bamfile(job, bamfile, sample_type, univ_options):
